spec: Make commit field optional
The buildrequest and testrequest messages carry a commit property that
is used to specify the exact commit that should be tested. However, it
is not always desirable or possible to specify the exact commit, so
the property should be optional.
This commit removes the property from the list of required properties,
allowing senders to omit the property.