spec: Remove "logs" property from test and build messages
The test and build messages have a required property that the sender
must use to send the names of generated logfiles. However, since
processes will generally register log messages with the context, it
is not necessary to send the names of logfiles in an IPC message.
This commit removes the "logs" property from the specification of the
test and build messages.