include/msg/merge: Remove log property
The log property was removed from the message definition of merge
messages, but the foundry/msg/merge module provides means to set and
get the value of this property.
This commit removes the log argument from foundry_msg_merge_new() and
the foundry_msg_merge_get_log() method, so that the log property
cannot be accessed anymore.