buildbot: Avoid dpkg-buildpackage placing results in ".."
There is no commandline option to tell dpkg-buildpackage to place the
build artifacts in a directory other than "..". Since buildbot clones
the sources directly into a temporary directory, this causes the build
artifacts to be written to a path outside the temporary directory.
This commit modifies buildbot to clone the sources into a subdirectory
of the temporary directory so that no build artifacts will be placed
outside of the temporary build directory.