publish_test: Add script for publishing test results
There is currently no way to manually publish a test message, making
it necessary to run a test on a testbot instance in order to test the
behavior of dispatchbot.
This commit adds a script that allows the caller to publish a test
message without the need to run a testbot instance.