buildbot: Build from commit messages instead of build requests
Buildbot currently expects to receive explicit build orders in the
form of build requests. This makes it necessary to have a central
instance in the build system that sends such build requests.
This commit modifies buildbot to treat commit messages on certain
branches as build triggers, allowing buildbot to act autonomously.