include/msg/dist: Add information about sources to dist message
Dist messages do not carry any information about the sources used
to build the artifacts, making it needlessly complicated to trace
the build/distribution.
This commit adds fields for the repository, branch, and commit that
the artifacts were built from. Further, this commit renames the
field for the package repository to "distribution" so it won't be
confused with the source repository.