distbot: Distribute packages upon receipt of sign messages
Distbots only distribute packages if they have been explicitly
instructed to do so by a dist request message. This makes it
necessary to have a central process in the build system that sends
such messages.
This commit modifies distbot to distribute packages upon receipt of
sign messages, eliminating the need to have a central orchestrator
in the build system.