buildbot: Check out the target branch in a separate statement
Buildbot fails to perform any builds because it attempts to check out
a ref (referencing a commit) using git clone's `-b' switch, which only
accepts branch references.
This commit modifies buildbot to check out the ref in a separate call
to `git checkout', which allows commit refs to be checked out.