buildbot: Prepend an entry to the changelog of unstable builds
Unstable branches are built automatically, but the changelog is not
changed everytime the unstable branch is updated, causing the built
packages to have the same version and release.

This commit modifies buildbot so that it prepends an entry to the
debian changelog of unstable builds, causing the package version and
release to be increased.