Remove unused bots and scripts
There are several bots and scripts left over from the old foundry
message architecture. Since the new message architecture does not
use command messages, scripts for sending commands cannot be used
anymore.
Further, mergebot and testbot do not support the new architecture,
and dispatchbot is not necessary with the new architecture.

This commit removes the unusable bots and command scripts.