Add COPYING file and GPLv3 license notice to source files
There is no indication in the source tree, what license the sources
are published under, making it unclear if and how the code may be used.

This commit adds a COPYING file with the full GPLv3 license text to the
root directory of the source tree, and a license notice to all source
files.