src/watchbot.sh: Don't notify if a HEAD could not be fetched
When checking if a branch has been changed, watchbot does not check
the return value of `fetch_head()`, the function that fetches the
hash of the latest commit. Because of this, watchbot compares the
old hash to an invalid value and erroneously sends a notification.

This commit modifies watchbot to check if the hash that was fetched
by `fetch_head()` is invalid before comparing it to the old hash.