foundry/context: Add function to get the context root directory
The context root directory is kept in the foundry/context module in
a global/private variable that users are discouraged to access.
However, to implement the transfer of contexts using toolbox's
ipc-tap and ipc-inject, it is necessary for users to query the
directory where contexts are stored.

This commit adds the foundry_context_get_root() function, which
returns the storage location of foundry contexts.