src/foundry-ctx{tap,inject}: Add context transfer hooks
Toolbox's ipc-tap and ipc-inject can be used to bridge IPC messages
between two foundry instances over SSH. However, foundry contexts
are not exchanged using IPC messages, which makes it necessary to
synchronize them separately whenever a build or sign message was
received.

This commit adds the foundry-ctxtap and foundry-ctxinject scripts,
which are hooks for ipc-tap and ipc-inject that synchronize contexts
whenever a build or sign message was published on a user-specified
topic.