foundry: Add crude configuration and start/stop-script
Foundry has neither a configuration nor a script for starting and
stopping the build system, making it necessary to start all of its
components manually. This makes starting and stopping foundry kind
of annoying and error-prone.

This commit adds a very simple Bash-based configuration file and a
script that can be used to start/stop/restart the build system.