Allow user to choose ipc flavor, default to uipc
Foundry uses the ipc module for IPC messaging, which requires GPG
to be correctly configured, which presents a hurdle for new users.
Since the ipc and uipc modules are interchangable, the user should
be able to decide what module to use.

This commit modifies all bots so that the user may select the ipc
flavor using the new `--proto' command line option. If the option
is not specified, uipc is used.
Further, foundry-ctxtap is changed to be protocol-independent,
meaning that it expects to receive foundry_msg objects instead of
ipc_msg objects on stdin.