bots/distbot: Avoid return value masking when checking signatures
When distbot checks the signature of a package, it pipes the output
from `dpkg-sig' to `log_info()' and checks the return value of the
entire pipe. Because `log_info()' almost always returns zero, the
pipe is highly unlikely to fail, making distbot unable to detect
invalid signatures.

This commit modifies distbot so that it checks the return value of
`dpkg-sig' directly.