bots/distbot: Allow branch->codename mapping to be configured
Distbot does not allow the caller to configure under which codename
a package is distributed, leaving the user no choice but to use the
default mapping.

This commit modifies distbot so that the mapping from branches to
codenames can be configured using the `--codename' command line
option. Mappings can be specified in two ways:

    --codename main:unstable
    --codename stable

The first instructs distbot to put packages built from the `main'
branch into the reposiory with codename `unstable'. The second form
declares the codename `stable' as a catch-all for all branches that
don't have an explicit mapping.