bots/buildbot: Require source code to be signed
Buildbot does not check whether the source code is signed, allowing
source code of an uncertain origin to be built and end up in package
repositories.

This commit modifies buildbot so that it checks the signature of the
commit it is going to build from. If the commit does not have a valid
signature and buildbot was invoked without `-U' / `--allow-unsigned',
it will refuse to build.