foundry/vargs: Implement module for variable argument lists
Because shell uses positional parameters to pass data to functions,
the position of a variable in the argument list is significant. This
makes calls to functions with a lot of arguments error prone.

This commit adds a module for variable argument lists. When used by
a function, positional parameters of the function must be preceeded
by an argument name, allowing the caller to pass the parameters in
an arbitrary order.