bots/yumwatchbot: Add bot for monitoring YUM repositories
To build packages for a distribution like MIRACLE LINUX that is
based on another RPM-based distribution, it is necessary to have
a bot in the build system that watches an upstream YUM repository
for changes.

This commit adds yumwatchbot, which periodically checks YUM
repositories for new or modified packages and sends msgv2 messages
if a new or modified package has been detected.