include/foundry: Add sourcepub type
There are no functions that allow a caller to access the sourcerefs
contained in a foundry.msg.source.new message, making it necessary
that the user manipulates the messages directly.

This commit adds the foundry/sourcepub module that provides methods
to access the data in a source.new messages, as well as a method for
the foundry/msgv2 that can be used to get the sourcepub from a
source.new message.