foundry/sourceref: Add method for retrieving the uri
Sourceref objects of all types contain a uri that points to the
source that is referenced in the object. However, the sourceref
module does not have a method to retrieve the uri field.

This commit adds the `foundry_sourceref_get_uri()' method that
allows the caller to conveniently retrieve a sourceref's uri.