foundry/sourcemod: Add sourcemod type for source.modified messages
The `foundry_msgv2_new()' method cannot create source.modified
messages because there is no type to handle the contents of such
messages.

This commit adds the sourcemod type to handle the contents of
source.modified messages, and modifies the `foundry_msgv2_new()'
method so that it can create source.modified messages.