foundry/msgv2: Make foundry_msgv2_is_type() check multiple types
The function `foundry_msgv2_is_type()` checks only one type at a
time, making it necessary to call the function multiple times to
check if a message is one out of multiple types, causing overhead.

This commit modifies `foundry_msgv2_is_type()` so that it can be
used to check if a message is one of multiple types.