foundry/checksum: Add functino to make checksum objects from files
There is no function to conveniently generate checksum objects from
files in the file system. Because of this, users have to compute
file hashes manually and pass them to `foundry_checksum_new()`.

This commit adds the `foundry_checksum_new_from_file()` function,
which generates a checksum object based on the algorithm and file
path passed by the caller.