foundry/directory: Add functions for making directory listings
Because there are no functions for listing the files and directories
in a directory object, the user has to manually access the data
inside the object, which is inconvenient and error-prone.

This commit adds the `foundry_directory_listing()` function, which
creates a listing of all files and directories in a directory.