foundry/timestamp: Add functions for timestamp creation and conversion
Because there are no functions for creating timestamps, foundry code has
to call `date` directly. Thus different parts of the code base might end
up invoking `date` with different parameters, causing inconsistencies.

This commit adds a module with functions for generating timestamps in a
unified format, and for converting timestamps to and from unix time.