foundry/buildresult: Add module to handle groups of builds
When a process generates a build result message, it has to store
build objects for each performed build inside a buildresult object,
which is then wrapped in a message and published. However, because
there are no functions for handling buildresult objects, components
have to provide their own implementation.

This commit adds a module with functions to generate buildresult
objects and retrieve the data that they contain.