bots/orderbot: Add build scheduler for RPM rebuilds
To enable rebuilds of RPM packages for distributions such as MIRACLE
LINUX or AlmaLinux, foundry needs a bot that published build request.

This commit adds orderbot, which listens for source.new messages and
sends build.request messages for each SRPM for the distribution that
was specified on the command line.