bots/spinbot: Track memory and disk usage of build process
Because spinbot does not implement resource tracking of mock builds,
there is no simple way for a user to find out how muck disk space or
memory was used to build a package.

This commit modifies spinbot so that it polls the memory and disk
usage of the build process, and includes that information in the
build result message.
The memory usage is estimated by summing up the memory use of the
build process and all sub-processes. The disk usage reflects the
size of the build root excluding any caches that mock may be using.