foundry/common: Fix incorrect check if endpoint should be closed
When `foundry_common_make_endpoint()` fails to subscribe the new
endpoint to one or more topics, it should clean up the endpoint
only if it is not shared with other processes. However, the check
whether the endpoint is a shared one is incorrect, causing the
endpoint to be cleaned up only when it is shared with other
processes.

This commit reverses the check in `foundry_common_make_endpoint()`
so that the endpoint is closed only when it does not have a name,
meaning that it is anonymous and not shared.