foundry/common: Do not print endpoint name when subscribe fails
The `foundry_common_make_endpoint()` function prints the name of
the endpoint even when the subscribe failed and the endpoint is
closed. The endpoint name should be printed only when the endpoint
was successfully created.

This commit modifies `foundry_common_make_endpoint()` so that it
prints the name of the endpoint only when it was successful.