bots/buildbot: Add support for custom repositories
Buildbot does not support custom repositories, making it unable to
build packages that depend on packages that are not in the official
repositories.

To allow building of packages that are not in the official
repositories, add the `--repository` and `--keyring` options which
can be used to add a custom repository and a GPG keyring used to
verify packages in a repository, respectively.
Repositories must be specified as space-separated string containing
the URL, suite, and one or more components (that is, the same way
they are listed in /etc/apt/sources.list but without the `deb`).
Keyrings must be either absolute or relative paths to GPG keyrings
in the local filesystem.

Attention: HTTPS repositories are not supported yet.